+++ src/Currency.php
@@ -59,7 +59,7 @@ final readonly class Currency implements Stringable, JsonSerializable
      */
     public static function of(string $currencyCode): Currency
     {
-        return ISOCurrencyProvider::getInstance()->getCurrency($currencyCode);
+        return IsoCurrencyProvider::getInstance()->getCurrency($currencyCode);
     }
 
     /**
@@ -73,7 +73,7 @@ final readonly class Currency implements Stringable, JsonSerializable
      */
     public static function ofCountry(string $countryCode): Currency
     {
-        return ISOCurrencyProvider::getInstance()->getCurrencyForCountry($countryCode);
+        return IsoCurrencyProvider::getInstance()->getCurrencyForCountry($countryCode);
     }
 
     /**
@@ -90,7 +90,7 @@ final readonly class Currency implements Stringable, JsonSerializable
      */
     public static function ofNumericCode(int $currencyCode): Currency
     {
-        return ISOCurrencyProvider::getInstance()->getCurrencyByNumericCode($currencyCode);
+        return IsoCurrencyProvider::getInstance()->getCurrencyByNumericCode($currencyCode);
     }
 
     /**